requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.2.1
fastapi==0.115.12
h11==0.16.0
idna==3.10
pydantic==2.11.6
pydantic_core==2.33.2
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.34.3

python-dotenv==1.0.1
pydantic[email]==2.11.6
motor==3.4.0
passlib[bcrypt]==1.7.4
python-jose==3.3.0

onnxruntime 
transformers==4.44.2
faster_whisper==1.0.3
gradio==4.19.2 
gradio-client==0.10.1
spacy==3.7.5
coqui-tts[languages]==0.24.2
cutlet==0.5.0
fugashi[unidic-lite]==1.4.0
torch==2.1.2
torchvision==0.16.2
torchaudio==2.1.2
